backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
bcrypt
python-dotenv
httpx
groq
PyJWT
python-multipart
